Completed human detection module and added human detection to be part of the control loop.

+++ DroneController/packages/madara_client/human_detection/BasicStrategy.h
@@ -44,7 +44,7 @@ namespace SMASH
          * @param min Minimum ambient temperature observed.
          * @param max Maximum ambient temperature observed.
          **/
-        void calculate_ambient_temp (int& min, int& max);
+        void calculate_ambient_temp (double& min, double& max);
     };  
   }
 }
